relocate sandbox verification scripts, update path lookup for sandbox.js file
