+++ packages/angular-playground/cli/src/check-errors/reporters/json-reporter.ts
