+++ packages/angular-playground/cli/src/run.ts
@@ -1,8 +1,8 @@
 import { configure, Config } from './configure';
 import { buildSandboxes } from './build-sandboxes';
 import { startWatch } from './start-watch';
-import { verifySandboxes } from './verify-sandboxes';
-import { findFirstFreePort } from './find-port';
+import { verifySandboxes } from './check-errors/verify-sandboxes';
+import { findFirstFreePort } from './check-errors/find-port';
 import { serveAngularCli } from './serve-angular-cli';
 import { buildAngularCli } from './build-angular-cli';
